Move create tests to a separate file
